+++ src/client/connector.rs
@@ -171,6 +171,12 @@ impl Node {
                     Err(e) => Err(NodeError::SessionError(self.address.clone(), e)),
                 }
             }
+            Command::Start { container_id } => {
+                match command::run_start(&self.address, session, sudo, &container_id).await {
+                    Ok(result) => Ok(result),
+                    Err(e) => Err(NodeError::SessionError(self.address.clone(), e)),
+                }
+            }
             Command::Stop { container_id } => {
                 match command::run_stop(&self.address, session, sudo, &container_id).await {
                     Ok(result) => Ok(result),