+++ src/utility/run.rs
@@ -36,7 +36,7 @@ pub async fn run_command(
             workdir,
         } => {
             let node_containers: Vec<(String, String)> =
-                find_container(client, &container_id, sudo, identity_file).await;
+                find_container(client, &container_id, sudo, false, identity_file).await;
 
             match node_containers.len() {
                 0 => {
@@ -122,7 +122,7 @@ pub async fn run_command(
             until,
         } => {
             let node_containers: Vec<(String, String)> =
-                find_container(client, &container_id, sudo, identity_file).await;
+                find_container(client, &container_id, sudo, false, identity_file).await;
             match node_containers.len() {
                 0 => {
                     vec![Err(CommandError::NoNodesFound(container_id))]
@@ -200,7 +200,7 @@ pub async fn run_command(
         }
         Command::Restart { time, container_id } => {
             let node_containers: Vec<(String, String)> =
-                find_container(client, &container_id, sudo, identity_file).await;
+                find_container(client, &container_id, sudo, true, identity_file).await;
 
             match node_containers.len() {
                 0 => {
@@ -227,9 +227,38 @@ pub async fn run_command(
                 }
             }
         }
+        Command::Start { container_id } => {
+            let node_containers: Vec<(String, String)> =
+                find_container(client, &container_id, sudo, true, identity_file).await;
+
+            match node_containers.len() {
+                0 => {
+                    vec![Err(CommandError::NoNodesFound(container_id))]
+                }
+                1 => {
+                    // unwrap is safe here since we .unwrap()check if there is exactly 1 element
+                    let node_tuple = node_containers.get(0).unwrap().to_owned();
+                    let node = Node::new(node_tuple.1);
+                    match node
+                        .run_command(Command::Start { container_id }, sudo, identity_file)
+                        .await
+                    {
+                        Ok(s) => vec![Ok(s)],
+                        Err(e) => vec![Err(CommandError::NodeError(e))],
+                    }
+                }
+                _ => {
+                    let nodes = node_containers
+                        .iter()
+                        .map(|(_, result)| result.clone())
+                        .collect::<Vec<String>>();
+                    vec![Err(CommandError::MutlipleNodesFound(nodes))]
+                }
+            }
+        }
         Command::Stop { container_id } => {
             let node_containers: Vec<(String, String)> =
-                find_container(client, &container_id, sudo, identity_file).await;
+                find_container(client, &container_id, sudo, false, identity_file).await;
 
             match node_containers.len() {
                 0 => {