+++ CHANGELOG.md
@@ -4,7 +4,9 @@ v0.0.6
 ## Added
 
 [[#33](https://github.com/MitchellBerend/docker-manager/pull/33)] Added a formatter for ps command
+
 [[#26](https://github.com/MitchellBerend/docker-manager/issues/26)] Added the restart command
+
 
 v0.0.5
 ---
@@ -12,7 +14,9 @@ v0.0.5
 ## Added
 
 [[#21](https://github.com/MitchellBerend/docker-manager/issues/21)] Remove implicit sudo call
+
 [[#12](https://github.com/MitchellBerend/docker-manager/issues/12)] Added a flag to filter nodes
+
 
 v0.0.4
 ---
@@ -20,8 +24,11 @@ v0.0.4
 ## Added
 
 [[#11](https://github.com/MitchellBerend/docker-manager/issues/11)] Implement images command and all its flags
+
 [[#17](https://github.com/MitchellBerend/docker-manager/pull/17)] Added shell completion generation
+
 [[#18](https://github.com/MitchellBerend/docker-manager/pull/18)] Added readme and license
+
 
 v0.0.3
 ---
@@ -29,7 +36,9 @@ v0.0.3
 ## Added
 
 [[#13](https://github.com/MitchellBerend/docker-manager/issues/13)] Implement exec command and most flags
+
 [[#15](https://github.com/MitchellBerend/docker-manager/pull/15)] Added bash compeltion command
+
 
 
 v0.0.2
@@ -38,4 +47,6 @@ v0.0.2
 ## Added
 
 [[#5](https://github.com/MitchellBerend/docker-manager/issues/5)] Implementremaining logs flags
+
 [[#6](https://github.com/MitchellBerend/docker-manager/issues/6)] Implement remaining ps flags
+