+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+
+
+[[#43](https://github.com/MitchellBerend/docker-manager/issues/43)] Added missing system command and flag descriptions
+
 [[#41]](https://github.com/MitchellBerend/docker-manager/pull/41) Refactor of formatter struct