+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+[[#52]](https://github.com/MitchellBerend/docker-manager/pull/52) Add a identity file flag option
+
 [[#45]](https://github.com/MitchellBerend/docker-manager/pull/45) Add formatterfor images command
 
 [[#43](https://github.com/MitchellBerend/docker-manager/issues/43)] Added missing system command and flag descriptions